Refactor dashboard frontend for clarity
