--- graph_data/influx/assets/style.css
@@ -1,80 +1,0 @@
-body {
-    margin: 0;
-    padding: 0;
-    background-color: black !important;
-    color: white;
-    font-family: Arial, sans-serif;
-}
-
-.app-container {
-    background-color: black;
-    color: white;
-    padding: 20px;
-    height: 100vh;
-    box-sizing: border-box;
-}
-
-.app-title {
-    text-align: center;
-    color: white;
-    margin-bottom: 10px;
-    font-size: 28px;
-}
-
-.last-updated {
-    text-align: left;
-    margin-bottom: 20px;
-    font-size: 14px;
-    color: #ccc;
-}
-/* Center column header text in AG Grid */
-.ag-theme-alpine-dark .ag-header-cell-label {
-    justify-content: center !important;
-}
-/* Center group column headers horizontally */
-.ag-theme-alpine-dark .ag-header-group-cell-label {
-    display: flex !important;
-    justify-content: center !important;
-    align-items: center !important;
-    text-align: center !important;
-}
-/* Force group header text to take full width
-.ag-theme-alpine-dark .ag-header-group-cell-label > .ag-header-group-text {
-    width: 100%;
-    text-align: center;
-} */
-
-/* Force black background for dropdown control */
-.dark-dropdown .Select__control,
-.dark-dropdown .Select-control {
-    background-color: black !important;
-    border: 1px solid #555 !important;
-    color: white !important;
-}
-
-/* Input text */
-.dark-dropdown .Select__input,
-.dark-dropdown .Select-input input {
-    color: white !important;
-    background-color: black !important;
-}
-
-/* Selected values text */
-.dark-dropdown .Select__single-value,
-.dark-dropdown .Select-value-label {
-    color: white !important;
-}
-
-/* Menu background */
-.dark-dropdown .Select__menu,
-.dark-dropdown .Select-menu-outer {
-    background-color: black !important;
-    color: white !important;
-}
-
-/* Hover effect */
-.dark-dropdown .Select__option--is-focused,
-.dark-dropdown .Select-option.is-focused {
-    background-color: #333 !important;
-    color: white !important;
-}
