--- graph_data/influx/png_graph/vali_data.csv
@@ -1,3 +1,0 @@
-validator_uid,run_id,epoch,participating_miners
-25,6,1,17
-25,6,2,26
